数据备份和恢复如果对端服务部存在给与提示

+++ tools/back-store.sh
@@ -66,26 +66,31 @@ function fdfs_back(){
     echo "fastDfs数据库开始备份,服务器IP:${from_fdfs_ip}"
     mkdir -p ${local_back_dir}/from
     fdfs_storage_file=${local_back_dir}/from/storage.conf
-    sshpass -p ${from_password}  scp  -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null root@${from_fdfs_ip}:/etc/fdfs/storage.conf ${fdfs_storage_file}
-    dirs_result=`cat ${fdfs_storage_file} | grep store_path0 |awk -F = '{printf $2}'`
-    dirs=(${dirs_result//\// })
-    max_dir=${dirs[0]}
-    if [[ ${max_dir} == "ruijie" ]];then
-		max_dir=""
-    else
-		max_dir="/${max_dir}"
-    fi
+	sshpass -p ${from_password}  scp  -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null root@${from_fdfs_ip}:/etc/fdfs/storage.conf ${fdfs_storage_file}
+	if [  -f "$fdfs_storage_file" ]; then
+		dirs_result=`cat ${fdfs_storage_file} | grep store_path0 |awk -F = '{printf $2}'`
+		dirs=(${dirs_result//\// })
+		max_dir=${dirs[0]}
+		if [[ ${max_dir} == "ruijie" ]];then
+			max_dir=""
+		else
+			max_dir="/${max_dir}"
+		fi
 
-    remote_back_dir="${max_dir}/${from_append_dir}"
-    del_command="rm -rf ${remote_back_dir}}/fdfs_data_back.tar > /dev/null 2>&1"
-    dir_create="mkdir -p ${remote_back_dir}";
-    back_command="cd ${dirs_result}/data; tar  -zcvf ${remote_back_dir}/fdfs_data_back.tar *> /dev/null 2>&1";
-    remote_ssh_command ${from_fdfs_ip} ${from_fdfs_password} "${del_command};${dir_create};${back_command};" 
-    ## 远程拷贝到本地目录，然后清空远程备份
-    rm -r ${local_back_dir}/fdfs_data_back.tar  > /dev/null 2>&1
-    mkdir -p ${local_back_dir} 
-    sshpass -p ${from_fdfs_password}  scp  -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null root@${from_fdfs_ip}:${remote_back_dir}/fdfs_data_back.tar ${local_back_dir}
-    remote_ssh_command ${from_fdfs_ip} ${from_fdfs_password} "${del_command};" 
+		remote_back_dir="${max_dir}/${from_append_dir}"
+		del_command="rm -rf ${remote_back_dir}}/fdfs_data_back.tar > /dev/null 2>&1"
+		dir_create="mkdir -p ${remote_back_dir}";
+		back_command="cd ${dirs_result}/data; tar  -zcvf ${remote_back_dir}/fdfs_data_back.tar *> /dev/null 2>&1";
+		remote_ssh_command ${from_fdfs_ip} ${from_fdfs_password} "${del_command};${dir_create};${back_command};" 
+		## 远程拷贝到本地目录，然后清空远程备份
+		rm -r ${local_back_dir}/fdfs_data_back.tar  > /dev/null 2>&1
+		mkdir -p ${local_back_dir} 
+		sshpass -p ${from_fdfs_password}  scp  -o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null root@${from_fdfs_ip}:${remote_back_dir}/fdfs_data_back.tar ${local_back_dir}
+		remote_ssh_command ${from_fdfs_ip} ${from_fdfs_password} "${del_command};" 
+	else
+	     echo "fdfs未配置，不进行备份,服务器IP:${to_fdfs_ip},配置文件:${fdfs_storage_file}"
+	fi
+
  fi
 }
 
@@ -101,7 +106,7 @@ function mongo_back(){
 
  	del_command="rm -rf ${remote_back_dir}/mongodb* > /dev/null 2>&1"
     dir_create="mkdir -p ${remote_back_dir}/mongodb";
-	back_command="mongodump -h ${from_mongo_ip} -u ${from_mongo_user} -p ${from_mongo_password} --authenticationDatabase 'admin'  -o ${remote_back_dir}/mongodb > /dev/null 2>&1"
+	back_command="mongodump -h ${from_mongo_ip} -u ${from_mongo_user} -p ${from_mongo_password} --authenticationDatabase 'admin'  -o ${remote_back_dir}/mongodb"
 	tar_command="cd ${remote_back_dir}/mongodb;tar  -zcvf ${remote_back_dir}/mongodb.tar * > /dev/null 2>&1";
 	remote_ssh_command ${from_mongo_ip} ${from_password} "${del_command};${dir_create};${back_command};${tar_command};"
